fix(cli): address reachability review feedback
