+++ apps/cli/src/config/validation.ts
@@ -8,6 +8,7 @@ import type {
   TokenPricingUsdPerMillion,
 } from './types.js';
 import { validateServiceMetadata } from './service-metadata.js';
+import { parseHostPort } from './public-address.js';
 
 const SERVICE_CATEGORY_PATTERN = /^[a-z0-9][a-z0-9-]*$/;
 const MAX_PUBLIC_ADDRESS_LENGTH = 255;
@@ -122,24 +123,7 @@ function parsePublicAddress(value: string): { host: string; port: number } | nul
   if (trimmed.length === 0 || trimmed.length > MAX_PUBLIC_ADDRESS_LENGTH) {
     return null;
   }
-
-  const lastColon = trimmed.lastIndexOf(':');
-  if (lastColon <= 0 || lastColon === trimmed.length - 1) {
-    return null;
-  }
-
-  const host = trimmed.slice(0, lastColon).trim();
-  const portText = trimmed.slice(lastColon + 1);
-  if (!/^\d+$/.test(portText)) {
-    return null;
-  }
-
-  const port = Number(portText);
-  if (host.length === 0 || !Number.isInteger(port) || port < 1 || port > 65535) {
-    return null;
-  }
-
-  return { host, port };
+  return parseHostPort(trimmed);
 }
 
 function isValidDomainName(value: string): boolean {
